git-svn-id: svn://svn.code.sf.net/p/kactus2/code/trunk@1043 f0aa6db5-7f46-43cc-a8a5-a31efdaafa67

+++ Vhdl2IPXact/vhdlanalysis.cpp
@@ -531,7 +531,7 @@ void VHDLanalysis::analyzeData(QTableWidget& generics_t,QTableWidget& ports_t, Q
             //determine width of integer
             if(type == "integer")
             {
-                width = (int)ceil(log(max)/log(2));
+                width = (int)ceil(log((double)max)/log((double)2));
             }
             //width in other types
             else if(type == "std_logic_vector")
